# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Template to use in your challenge",
        "title": "Core API template",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/product": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a paginated list of products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get products paginated",
                "parameters": [
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "x-enum-varnames": [
                            "Asc",
                            "Desc"
                        ],
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "created_at_geq",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "created_at_leq",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The unique identifier",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "ingredients",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "The offset from where to start the items",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "name": "price",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "x-enum-varnames": [
                            "Asc",
                            "Desc"
                        ],
                        "name": "updated_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updated_at_geq",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updated_at_leq",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/product_entity.Product"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates a product's details by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update product by ID",
                "parameters": [
                    {
                        "description": "Product data to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product_model.UpdateWithId"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "schema": {
                            "$ref": "#/definitions/product_entity.Product"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new product item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Creates a new product",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product_model.Create"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created product",
                        "schema": {
                            "$ref": "#/definitions/product_entity.Product"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a product by its ID (only accessible by admins)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete product by ID",
                "parameters": [
                    {
                        "description": "Product ID to delete",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_cogniia_core-api-template_src_common_model.RequiredId"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_cogniia_core-api-template_src_common_model.RequiredId": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "The unique identifier.",
                    "type": "string"
                }
            }
        },
...
        "product_entity.Product": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "product_model.Create": {
            "type": "object",
            "properties": {
                "ingredients": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "product_model.UpdateWithId": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "The unique identifier.",
                    "type": "string"
                },
                "ingredients": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}